moved sync committee subscriptions for VC api
as part of moving validator api to using typedefs, moved the postSyncCommitteeSubscriptions.

there are still several more to move, but this is plenty for one PR.
