+++ data/beaconrestapi/src/main/java/tech/pegasys/teku/beaconrestapi/handlers/v1/validator/PostSyncCommitteeSubscriptions.java
@@ -13,52 +13,26 @@
 
 package tech.pegasys.teku.beaconrestapi.handlers.v1.validator;
 
+import static tech.pegasys.teku.ethereum.json.types.validator.PostSyncCommitteeData.SYNC_COMMITTEE_SUBSCRIPTION;
 import static tech.pegasys.teku.infrastructure.http.HttpStatusCodes.SC_OK;
 import static tech.pegasys.teku.infrastructure.http.RestApiConstants.TAG_VALIDATOR;
 import static tech.pegasys.teku.infrastructure.http.RestApiConstants.TAG_VALIDATOR_REQUIRED;
-import static tech.pegasys.teku.infrastructure.json.types.CoreTypes.INTEGER_TYPE;
-import static tech.pegasys.teku.infrastructure.json.types.CoreTypes.UINT64_TYPE;
 
 import com.fasterxml.jackson.core.JsonProcessingException;
-import it.unimi.dsi.fastutil.ints.IntArrayList;
-import it.unimi.dsi.fastutil.ints.IntOpenHashSet;
-import it.unimi.dsi.fastutil.ints.IntSet;
 import java.util.List;
-import java.util.Objects;
 import tech.pegasys.teku.api.DataProvider;
 import tech.pegasys.teku.api.ValidatorDataProvider;
+import tech.pegasys.teku.ethereum.json.types.validator.PostSyncCommitteeData;
+import tech.pegasys.teku.ethereum.json.types.validator.SyncCommitteeSubnetSubscription;
 import tech.pegasys.teku.infrastructure.json.types.DeserializableTypeDefinition;
 import tech.pegasys.teku.infrastructure.restapi.endpoints.AsyncApiResponse;
 import tech.pegasys.teku.infrastructure.restapi.endpoints.EndpointMetadata;
 import tech.pegasys.teku.infrastructure.restapi.endpoints.RestApiEndpoint;
 import tech.pegasys.teku.infrastructure.restapi.endpoints.RestApiRequest;
-import tech.pegasys.teku.infrastructure.unsigned.UInt64;
-import tech.pegasys.teku.validator.api.SyncCommitteeSubnetSubscription;
 
 public class PostSyncCommitteeSubscriptions extends RestApiEndpoint {
   public static final String ROUTE = "/eth/v1/validator/sync_committee_subscriptions";
   private final ValidatorDataProvider provider;
-
-  static final DeserializableTypeDefinition<PostSyncCommitteeData> REQUEST_TYPE =
-      DeserializableTypeDefinition.object(PostSyncCommitteeData.class)
-          .name("PostSyncCommitteeData")
-          .initializer(PostSyncCommitteeData::new)
-          .withField(
-              "validator_index",
-              INTEGER_TYPE,
-              PostSyncCommitteeData::getValidatorIndex,
-              PostSyncCommitteeData::setValidatorIndex)
-          .withField(
-              "sync_committee_indices",
-              DeserializableTypeDefinition.listOf(INTEGER_TYPE),
-              PostSyncCommitteeData::getSyncCommitteeIndices,
-              PostSyncCommitteeData::setSyncCommitteeIndices)
-          .withField(
-              "until_epoch",
-              UINT64_TYPE,
-              PostSyncCommitteeData::getUntilEpoch,
-              PostSyncCommitteeData::setUntilEpoch)
-          .build();
 
   public PostSyncCommitteeSubscriptions(final DataProvider dataProvider) {
     this(dataProvider.getValidatorDataProvider());
@@ -74,7 +48,7 @@ public class PostSyncCommitteeSubscriptions extends RestApiEndpoint {
                     + "Sync committees are not present in phase0, but are required for Altair networks.\n\n"
                     + "Subscribing to sync committee subnets is an action performed by VC to enable network participation in Altair networks, and only required if the VC has an active validator in an active sync committee.")
             .tags(TAG_VALIDATOR, TAG_VALIDATOR_REQUIRED)
-            .requestBodyType(DeserializableTypeDefinition.listOf(REQUEST_TYPE))
+            .requestBodyType(DeserializableTypeDefinition.listOf(SYNC_COMMITTEE_SUBSCRIPTION))
             .response(SC_OK, "Successful response")
             .build());
     this.provider = validatorDataProvider;
@@ -89,79 +63,5 @@ public class PostSyncCommitteeSubscriptions extends RestApiEndpoint {
         provider
             .subscribeToSyncCommitteeSubnets(subscriptions)
             .thenApply(AsyncApiResponse::respondOk));
-  }
-
-  public static class PostSyncCommitteeData {
-    private int validatorIndex;
-    private IntSet syncCommitteeIndices;
-    private UInt64 untilEpoch;
-
-    public PostSyncCommitteeData() {}
-
-    public PostSyncCommitteeData(
-        final int validatorIndex, final IntSet syncCommitteeIndices, final UInt64 untilEpoch) {
-      this.validatorIndex = validatorIndex;
-      this.syncCommitteeIndices = syncCommitteeIndices;
-      this.untilEpoch = untilEpoch;
-    }
-
-    public SyncCommitteeSubnetSubscription toSyncCommitteeSubnetSubscription() {
-      return new SyncCommitteeSubnetSubscription(validatorIndex, syncCommitteeIndices, untilEpoch);
-    }
-
-    public int getValidatorIndex() {
-      return validatorIndex;
-    }
-
-    public void setValidatorIndex(final int validatorIndex) {
-      this.validatorIndex = validatorIndex;
-    }
-
-    public List<Integer> getSyncCommitteeIndices() {
-      return new IntArrayList(syncCommitteeIndices);
-    }
-
-    public void setSyncCommitteeIndices(final List<Integer> syncCommitteeIndices) {
-      this.syncCommitteeIndices = new IntOpenHashSet(syncCommitteeIndices);
-    }
-
-    public UInt64 getUntilEpoch() {
-      return untilEpoch;
-    }
-
-    public void setUntilEpoch(final UInt64 untilEpoch) {
-      this.untilEpoch = untilEpoch;
-    }
-
-    @Override
-    public boolean equals(final Object o) {
-      if (this == o) {
-        return true;
-      }
-      if (o == null || getClass() != o.getClass()) {
-        return false;
-      }
-      PostSyncCommitteeData that = (PostSyncCommitteeData) o;
-      return validatorIndex == that.validatorIndex
-          && Objects.equals(syncCommitteeIndices, that.syncCommitteeIndices)
-          && Objects.equals(untilEpoch, that.untilEpoch);
-    }
-
-    @Override
-    public int hashCode() {
-      return Objects.hash(validatorIndex, syncCommitteeIndices, untilEpoch);
-    }
-
-    @Override
-    public String toString() {
-      return "PostSyncCommitteeData{"
-          + "validatorIndex="
-          + validatorIndex
-          + ", syncCommitteeIndices="
-          + syncCommitteeIndices
-          + ", untilEpoch="
-          + untilEpoch
-          + '}';
-    }
   }
 }
